--- packages/client/src/utils/enums.ts
@@ -1,4 +1,0 @@
-export enum RedirectUrls {
-  VercelApp = 'https://among-us-client.vercel.app/signin',
-  YaPraktikumTech = 'https://enoty.ya-praktikum.tech:3000/signin',
-}
